# src/web/docs/swagger.yaml
basePath: /recruiter-api/v1
definitions:
...
  errorinterface.ErrorResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
...
  jobsinterface.JobRequest:
    properties:
      category:
        type: string
      createdTimestampUTC:
        type: string
      deleted:
        type: boolean
      employmentType:
        items:
          type: string
        type: array
      function:
        type: string
      goodToHaveSkills:
        items:
          type: string
        type: array
      id:
        type: string
      isRemote:
        type: boolean
      isVerified:
        type: boolean
      locations:
        items:
          type: string
        type: array
      maxExperience:
        type: integer
      minExperience:
        type: integer
      mustHaveSkills:
        items:
          type: string
        type: array
      organizationID:
        type: string
      recruiterID:
        type: string
      remoteTimezone:
        type: string
      sideNote:
        type: string
      summary:
        additionalProperties: true
        type: object
      title:
        type: string
      updatedTimestampUTC:
        type: string
      visaSponsorShip:
        type: boolean
    type: object
  jobsinterface.JobResponse:
    properties:
      category:
        type: string
      createdTimestampUTC:
        type: string
      deleted:
        type: boolean
      employmentType:
        items:
          type: string
        type: array
      function:
        type: string
      goodToHaveSkills:
        items:
          type: string
        type: array
      id:
        type: string
      isRemote:
        type: boolean
      isVerified:
        type: boolean
      locations:
        items:
          type: string
        type: array
      maxExperience:
        type: integer
      minExperience:
        type: integer
      mustHaveSkills:
        items:
          type: string
        type: array
      organizationID:
        type: string
      recruiterID:
        type: string
      remoteTimezone:
        type: string
      sideNote:
        type: string
      summary:
        additionalProperties: true
        type: object
      title:
        type: string
      updatedTimestampUTC:
        type: string
      visaSponsorShip:
        type: boolean
    type: object
host: localhost:8001
info:
  contact:
    email: hello@xyz.ai
    name: API Support
    url: http://xyz.ai
  description: Recruiter API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Recruiter API Documentation
  version: "2.0"
paths:
...
  /jobs/jobID:
    delete:
      consumes:
      - application/json
      description: It allows to delete a Job by ID
      parameters:
      - description: Job ID
        in: path
        name: jobID
        required: true
        type: string
      - description: BEARER JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Delete a Job
      tags:
      - jobs
    get:
      consumes:
      - application/json
      description: It allows to reterive a Job by ID
      parameters:
      - description: Job ID
        in: path
        name: jobID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jobsinterface.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Get a Job
      tags:
      - jobs
    put:
      consumes:
      - application/json
      description: It allows to update a Job
      parameters:
      - description: Job Details
        in: body
        name: '*'
        required: true
        schema:
          $ref: '#/definitions/jobsinterface.JobRequest'
      - description: BEARER JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Job ID
        in: path
        name: jobID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jobsinterface.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Update a  Job
      tags:
      - jobs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"